+++ android/app/src/main/kotlin/com/DaAppLab/DaKanjiRecognizer/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.DaAppLab.DaKanjiRecognizer
+package com.DaAppLab.DaKanji
 
 import io.flutter.embedding.android.FlutterActivity